:recycle: Refactored into PokemonList
